I finish it!

+++ CS321_W5D2_BlogAPI.Core/Services/PostService.cs
@@ -23,6 +23,17 @@ namespace CS321_W5D2_BlogAPI.Core.Services
             //     Use the _userService to get the current users id.
             //     You may have to retrieve the blog in order to check user id
             // TODO: assign the current date to DatePublished
+
+            var currentUserId = _userService.CurrentUserId;
+            var currentBlog = _blogRepository.Get(newPost.BlogId);
+
+            if (currentUserId != currentBlog.UserId)
+            {
+                throw new Exception("This is not your blog, please get permission or check your spelling");
+            }
+
+            newPost.DatePublished = DateTime.Now;
+
             return _postRepository.Add(newPost);
         }
 
@@ -45,12 +56,22 @@ namespace CS321_W5D2_BlogAPI.Core.Services
         {
             var post = this.Get(id);
             // TODO: prevent user from deleting from a blog that isn't theirs
+            var currentUserId = _userService.CurrentUserId;
+            if (post.Blog.UserId != currentUserId)
+            {
+                throw new Exception("This is not your blog, please get permission or check your spelling");
+            }
             _postRepository.Remove(id);
         }
 
         public Post Update(Post updatedPost)
         {
+            var currentUserId = _userService.CurrentUserId;
             // TODO: prevent user from updating a blog that isn't theirs
+            if (currentUserId != updatedPost.Blog.UserId)
+            {
+                throw new Exception("DENIED ACCESS! Bad person.");
+            }
             return _postRepository.Update(updatedPost);
         }
